refactor(jwk): add rcrypto type conversions for jose-jwk
