+++ jose-jwk/Cargo.toml
@@ -16,6 +16,12 @@ readme = "README.md"
 edition = "2021"
 rust-version = "1.63"
 
+[features]
+rcrypto-p256 = ["rcrypto", "dep:p256"]
+rcrypto-p384 = ["rcrypto", "dep:p384"]
+rcrypto-rsa = ["rcrypto", "dep:rsa"]
+rcrypto = []
+
 [dependencies]
 serde = { version = "1.0.136", default-features = false, features = ["alloc", "derive"] }
 jose-b64 = { path = "../jose-b64", default-features = false, features = ["secret"] }
@@ -23,6 +29,11 @@ jose-jwa = { path = "../jose-jwa" }
 
 # Optional Dependencies
 url = { version = "2.2.2", default-features = false, optional = true, features = ["serde"] }
+
+# Internal Dependencies
+p256 = { version = "0.11.1", default-features = false, optional = true, features = ["arithmetic"] }
+p384 = { version = "0.11.1", default-features = false, optional = true, features = ["arithmetic"] }
+rsa = { version = "0.7.0-rc.0", default-features = false, optional = true }
 
 [dev-dependencies]
 serde_json = "1.0.85"
