+++ jose-jwk/src/crypto/rcrypto/rsa.rs
@@ -1,14 +1,66 @@
 // SPDX-FileCopyrightText: 2022 Profian Inc. <opensource@profian.com>
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
-#![cfg(feature = "rsa")]
+#![cfg(feature = "rcrypto-rsa")]
 
 use rsa::{BigUint, PublicKeyParts, RsaPrivateKey, RsaPublicKey};
 
-use crate::jwk::{Rsa, RsaPrivate};
-use crate::key::rcrypto::Error;
+use jose_jwa::{Algorithm, Algorithm::Signing, Signing::*};
 
-#[cfg_attr(docsrs, doc(cfg(feature = "rsa")))]
+use super::super::KeyInfo;
+use super::Error;
+use crate::{Rsa, RsaPrivate};
+
+#[cfg_attr(docsrs, doc(cfg(feature = "rcrypto-rsa")))]
+impl KeyInfo for RsaPublicKey {
+    fn strength(&self) -> usize {
+        self.size() / 16
+    }
+
+    fn is_supported(&self, algo: &Algorithm) -> bool {
+        // RFC 7518 Section 3.3
+        //
+        // I would actually prefer stronger requirements here based on the
+        // algorithm below. However, the RFCs actually specify examples that
+        // this would break. Note that we generate stronger keys by default.
+        if self.strength() < 16 {
+            return false;
+        }
+
+        #[allow(clippy::match_like_matches_macro)]
+        match algo {
+            Signing(Rs256) => true,
+            Signing(Rs384) => true,
+            Signing(Rs512) => true,
+            Signing(Ps256) => true,
+            Signing(Ps384) => true,
+            Signing(Ps512) => true,
+            _ => false,
+        }
+    }
+}
+
+#[cfg_attr(docsrs, doc(cfg(feature = "rcrypto-rsa")))]
+impl KeyInfo for RsaPrivateKey {
+    fn strength(&self) -> usize {
+        self.size() / 16
+    }
+
+    fn is_supported(&self, algo: &Algorithm) -> bool {
+        #[allow(clippy::match_like_matches_macro)]
+        match (algo, self.strength()) {
+            (Signing(Rs256), 16..) => true,
+            (Signing(Rs384), 24..) => true,
+            (Signing(Rs512), 32..) => true,
+            (Signing(Ps256), 16..) => true,
+            (Signing(Ps384), 24..) => true,
+            (Signing(Ps512), 32..) => true,
+            _ => false,
+        }
+    }
+}
+
+#[cfg_attr(docsrs, doc(cfg(feature = "rcrypto-rsa")))]
 impl From<&RsaPublicKey> for Rsa {
     fn from(pk: &RsaPublicKey) -> Self {
         Self {
@@ -19,14 +71,14 @@ impl From<&RsaPublicKey> for Rsa {
     }
 }
 
-#[cfg_attr(docsrs, doc(cfg(feature = "rsa")))]
+#[cfg_attr(docsrs, doc(cfg(feature = "rcrypto-rsa")))]
 impl From<RsaPublicKey> for Rsa {
     fn from(sk: RsaPublicKey) -> Self {
         (&sk).into()
     }
 }
 
-#[cfg_attr(docsrs, doc(cfg(feature = "rsa")))]
+#[cfg_attr(docsrs, doc(cfg(feature = "rcrypto-rsa")))]
 impl TryFrom<&Rsa> for RsaPublicKey {
     type Error = Error;
 
@@ -37,7 +89,7 @@ impl TryFrom<&Rsa> for RsaPublicKey {
     }
 }
 
-#[cfg_attr(docsrs, doc(cfg(feature = "rsa")))]
+#[cfg_attr(docsrs, doc(cfg(feature = "rcrypto-rsa")))]
 impl TryFrom<Rsa> for RsaPublicKey {
     type Error = Error;
 
@@ -47,7 +99,7 @@ impl TryFrom<Rsa> for RsaPublicKey {
 }
 
 // TODO: patch rsa crate to export the optional values
-#[cfg_attr(docsrs, doc(cfg(feature = "rsa")))]
+#[cfg_attr(docsrs, doc(cfg(feature = "rcrypto-rsa")))]
 impl From<&RsaPrivateKey> for Rsa {
     fn from(pk: &RsaPrivateKey) -> Self {
         Self {
@@ -61,14 +113,14 @@ impl From<&RsaPrivateKey> for Rsa {
     }
 }
 
-#[cfg_attr(docsrs, doc(cfg(feature = "rsa")))]
+#[cfg_attr(docsrs, doc(cfg(feature = "rcrypto-rsa")))]
 impl From<RsaPrivateKey> for Rsa {
     fn from(sk: RsaPrivateKey) -> Self {
         (&sk).into()
     }
 }
 
-#[cfg_attr(docsrs, doc(cfg(feature = "rsa")))]
+#[cfg_attr(docsrs, doc(cfg(feature = "rcrypto-rsa")))]
 impl TryFrom<&Rsa> for RsaPrivateKey {
     type Error = Error;
 
@@ -84,7 +136,7 @@ impl TryFrom<&Rsa> for RsaPrivateKey {
                 let mut primes = alloc::vec![p, q];
                 primes.extend(opt.oth.iter().map(|x| BigUint::from_bytes_be(&x.r)));
 
-                return Ok(Self::from_components(n, e, d, primes));
+                return Self::from_components(n, e, d, primes).map_err(|_| Error::Invalid);
             }
 
             return Err(Error::Unsupported);
@@ -94,7 +146,7 @@ impl TryFrom<&Rsa> for RsaPrivateKey {
     }
 }
 
-#[cfg_attr(docsrs, doc(cfg(feature = "rsa")))]
+#[cfg_attr(docsrs, doc(cfg(feature = "rcrypto-rsa")))]
 impl TryFrom<Rsa> for RsaPrivateKey {
     type Error = Error;
