+++ jose/src/key/rcrypto/rsa.rs
@@ -29,60 +29,3 @@ impl<T: Type<Rsa = RsaPrivateKey>> From<RsaPrivateKey> for Key<T> {
         }
     }
 }
-
-#[cfg_attr(docsrs, doc(cfg(feature = "rsa")))]
-impl KeyInfo for RsaPublicKey {
-    fn strength(&self) -> usize {
-        self.size() / 16
-    }
-
-    fn is_supported(&self, algo: &Algorithm) -> bool {
-        // RFC 7518 Section 3.3
-        //
-        // I would actually prefer stronger requirements here based on the
-        // algorithm below. However, the RFCs actually specify examples that
-        // this would break. Note that we generate stronger keys by default.
-        if self.strength() < 16 {
-            return false;
-        }
-
-        match algo {
-            Signing(Rs256) => true,
-            Signing(Rs384) => true,
-            Signing(Rs512) => true,
-
-            #[cfg(feature = "rand")]
-            Signing(Ps256) => true,
-            #[cfg(feature = "rand")]
-            Signing(Ps384) => true,
-            #[cfg(feature = "rand")]
-            Signing(Ps512) => true,
-
-            _ => false,
-        }
-    }
-}
-
-#[cfg_attr(docsrs, doc(cfg(feature = "rsa")))]
-impl KeyInfo for RsaPrivateKey {
-    fn strength(&self) -> usize {
-        self.size() / 16
-    }
-
-    fn is_supported(&self, algo: &Algorithm) -> bool {
-        match (algo, self.strength()) {
-            (Signing(Rs256), 16..) => true,
-            (Signing(Rs384), 24..) => true,
-            (Signing(Rs512), 32..) => true,
-
-            #[cfg(feature = "rand")]
-            (Signing(Ps256), 16..) => true,
-            #[cfg(feature = "rand")]
-            (Signing(Ps384), 24..) => true,
-            #[cfg(feature = "rand")]
-            (Signing(Ps512), 32..) => true,
-
-            _ => false,
-        }
-    }
-}
